remove api folder, move files from api folder up one lvl to the routes directory
