--- apps/server/src/routes/api/index.js
@@ -1,8 +1,0 @@
-// const router = require('express').Router();
-// const userRoutes = require('./userRoutes');
-
-// router.use('/users', userRoutes);
-
-// module.exports = router;
-
-// move this code to server/src/main.js
