+++ apps/server/src/routes/index.js
@@ -1,12 +1,8 @@
 // const router = require('express').Router();
-// const apiRoutes = require('./api');
+// const userRoutes = require('./userRoutes');
 
-// router.use('/api', apiRoutes);
-
-// router.use((req, res) => {
-//   return res.send('Wrong route!');
-// });
+// router.use('/users', userRoutes);
 
 // module.exports = router;
 
-// moved this to server/src/main.js
+// move this code to server/src/main.js
